refactor(cloth-config): move classes back to `config.gui` package
Now that `:cloth-config` is bundled as part of the "main" mod, we don't
need to workaround FML's package conflict/encapsulation restrictions.
