+++ cloth-config/src/main/java/net/xolt/freecam/config/gui/ClothConfigScreenProvider.java
@@ -1,9 +1,7 @@
-package net.xolt.freecam.clothconfig;
+package net.xolt.freecam.config.gui;
 
 import net.minecraft.client.gui.screens.Screen;
 import net.xolt.freecam.config.controller.ConfigControllerRegistry;
-import net.xolt.freecam.config.gui.ConfigScreenProvider;
-import net.xolt.freecam.config.gui.OptionalProvider;
 import net.xolt.freecam.config.model.ModConfigDTO;
 import org.jetbrains.annotations.Nullable;
