+++ cloth-config/src/main/resources/META-INF/services/net.xolt.freecam.config.gui.ConfigScreenProvider
@@ -1,1 +1,1 @@
-net.xolt.freecam.clothconfig.ClothConfigScreenProvider
+net.xolt.freecam.config.gui.ClothConfigScreenProvider
